Team сделала апдейт фрагменту, вьюмоделька
+
правки для dog
