+++ app/src/main/java/ru/practicum/android/diploma/ui/team/TeamViewModel.kt
@@ -1,9 +1,9 @@
-package ru.practicum.android.diploma.ui.favourite
+package ru.practicum.android.diploma.ui.team
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class FavouriteViewModel : ViewModel() {
+class TeamViewModel : ViewModel() {
 
     val mutable = MutableLiveData<Int>()
